pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "langgraph-learn"
version = "0.1.0"
description = "A LangGraph learning project"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Qun Zheng", email = "stonezq88@gmail.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = [
    "langgraph",
    "langchain",
    "langchain-openai",
    "python-dotenv",
    "langsmith",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
jupyter = [
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "matplotlib>=3.5.0",
    "pandas>=1.5.0",
    "nest-asyncio>=1.5.0",
    "aiohttp>=3.8.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/langgraph-learn"
Repository = "https://github.com/yourusername/langgraph-learn"
Issues = "https://github.com/yourusername/langgraph-learn/issues"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
] 